entity 재정의, client 비활성화 작업

+++ pay-server/src/main/java/org/psawesome/payserver/domain/user/repo/PayUserRepository.java
@@ -9,6 +9,6 @@ import org.springframework.stereotype.Repository;
  * DATE: 20. 6. 26. Friday
  */
 @Repository
-public interface PayUserRepository extends ReactiveCrudRepository<PayUser, Long> {
+public interface PayUserRepository extends ReactiveCrudRepository<PayUser, Integer> {
 
 }
